feat: 예외 처리 핸들러 추가
